+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/UserMapper.java
@@ -13,7 +13,7 @@ public class UserMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         return User.builder()
-                .id(rs.getInt("user_id"))
+                .id(rs.getLong("user_id"))
                 .name(rs.getString("name"))
                 .email(rs.getString("email"))
                 .login(rs.getString("login"))
